Add files via upload
It is software used to manage data in library based on different operations. 
Features:-
It shows list of books available in library. 
It add books not available in library. 
It maintains records about which books are borrowed and returned by user.
Note:- Here a text file uploaded with main file as an input to be used while execution of the program.